Consolidate to mixin
